first draft of item creation + updating

+++ modules/lib/play-json-salat-utils/src/main/scala/org/corespring/play/json/salat/utils/converters.scala
@@ -136,7 +136,7 @@ object ToDBObject {
         case v: JsObject => parse(v, parent)
         case v: JsArray => parse(v, parent)
         case v: JsString => JsSuccess(v.value, parent)
-        case v: JsNumber => JsSuccess(v.value, parent)
+        case v: JsNumber => JsSuccess(v.value.toLong, parent)
         case v: JsBoolean => JsSuccess(v.value, parent)
         case JsNull => JsSuccess(null)
         case _: JsUndefined => JsSuccess(null)
